stuf

+++ src/offboard_py/scripts/tracker.py
@@ -134,7 +134,7 @@ class Tracker:
         self.logits = np.clip(self.logits, a_min = -2, a_max = 10)
 
         mask = np.logical_or(self.logits[..., 0] > 0, self.logits[..., 1] > 0)
-        mask = scale_image(mask, 0.5)
+        mask = scale_image(mask.astype(np.uint8), 0.5)
         print('-'*20)
         for c in range(mask.shape[1] -1, -1, -1):
             for r in range(mask.shape[0] -1, -1, -1):
